backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
asyncpg==0.29.0
aiosqlite==0.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
celery==5.3.6
redis==5.0.3
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
